Replace oc cluster up with minikube for PR jenkins jobs (#4098)
* Replace oc cluster up with minikube


* fixup! Replace oc cluster up with minikube


* fixup! fixup! Replace oc cluster up with minikube


* fixup! fixup! fixup! Replace oc cluster up with minikube


* fixup! fixup! fixup! fixup! Replace oc cluster up with minikube


* Set RBAC for all kube versions


* Allow network policy on minikube


* Revert changes for NP


* Set default version properly


* Set default versiopn fo result reporting from PR job to github
