+++ systemtest/scripts/results_info.sh
@@ -4,7 +4,7 @@ RESULTS_PATH=${1}
 TEST_CASE=${2}
 TEST_PROFILE=${3}
 BUILD_ID=${4:-0}
-OCP_VERSION=${5:-3}
+KUBE_VERSION=${5:-1.16.0}
 TEST_ONLY=${6:-''}
 EXCLUDED_GROUPS=${7:-''}
 ENV_VARIABLES=${8:-''}
@@ -31,12 +31,6 @@ TEST_ERRORS_COUNT=$(get_test_count "errors")
 TEST_SKIPPED_COUNT=$(get_test_count "skipped")
 TEST_FAILURES_COUNT=$(get_test_count "failures")
 
-if [[ "${OCP_VERSION}" == "4" ]]; then
-  BUILD_ENV="crc"
-else
-  BUILD_ENV="oc cluster up"
-fi
-
 TEST_ALL_FAILED_COUNT=$((TEST_ERRORS_COUNT + TEST_FAILURES_COUNT))
 
 if [[ -n "${EXCLUDED_GROUPS}" ]]; then
@@ -47,7 +41,7 @@ if [[ -n "${ENV_VARIABLES}" ]]; then
   ADDITIONAL_INFO+="**ENV_VARIABLES:** ${ENV_VARIABLES}\n"
 fi
 
-SUMMARY="**TEST_PROFILE**: ${TEST_PROFILE}\n${ADDITIONAL_INFO}**TEST_CASE:** ${TEST_CASE}\n**TOTAL:** ${TEST_COUNT}\n**PASS:** $((TEST_COUNT - TEST_ALL_FAILED_COUNT - TEST_SKIPPED_COUNT))\n**FAIL:** ${TEST_ALL_FAILED_COUNT}\n**SKIP:** ${TEST_SKIPPED_COUNT}\n**BUILD_NUMBER:** ${BUILD_ID}\n**BUILD_ENV:** ${BUILD_ENV}\n"
+SUMMARY="**TEST_PROFILE**: ${TEST_PROFILE}\n${ADDITIONAL_INFO}**TEST_CASE:** ${TEST_CASE}\n**TOTAL:** ${TEST_COUNT}\n**PASS:** $((TEST_COUNT - TEST_ALL_FAILED_COUNT - TEST_SKIPPED_COUNT))\n**FAIL:** ${TEST_ALL_FAILED_COUNT}\n**SKIP:** ${TEST_SKIPPED_COUNT}\n**BUILD_NUMBER:** ${BUILD_ID}\n**KUBE_VERSION:** ${KUBE_VERSION}\n"
 
 
 FAILED_TESTS=$(find "${RESULTS_PATH}" -name 'TEST*.xml' -type f -print0 | xargs -0 awk '/<testcase.*>/{ getline x; if (x ~ "<error" || x ~ "<failure") {  gsub(/classname=|name=|\"/, "", $0); if ($3 ~ "time=") { print "\\n- " $2 } else { print "\\n- " $2 " in " $3 } }}')
